You can now move around in the game and drop bombs.
AND the bombs blow up into a fire

BUT the fire does not actually affect the players and the players can walk straight through the bombs.

Need to decide if players will be allowed to move through the bombs and eachothers.

It is much more obvious thouhg that the players will be killed by the fire, so will have to decide how to handle that.